🔨 [projects] Move fixtures from `conftest.py` to `test_create`
